Predicate the call to trace_mem()
The pin sampler inserts a call to trace_mem() and a predicated
call to trace_update_access_counter() that is NOT called if the current instruction is a predicate (cmov, ...)
and the condition is false.

This could lead to the situation where the access counter is not updated but the memory access is recorded by the sampler, and
thus to the situation where two samples accessing the same memory location have the same timestamp.
This causes the current model to exit with this error:

  File "/home/moncef/coding/uart/util/python/util/analysis_orig/utils.py", line 50, in usf_read_events
    assert(event.begin.time < event.end.time)
AssertionError

This patch makes the call to trace_mem() predicated too so that if the current instruction is a predicate and the condition is false,
the memory access is not passed to the sampler